Consolidate dueDate, PlaylistTitle, and ProgressBar components into PlaylistHeader
