--- src/components/DueDate/DueDate.js
@@ -1,5 +1,0 @@
-import { P } from './styles'
-
-const DueDate = ({ dueDate }) => <P>Due: {dueDate}</P>;
-
-export default DueDate;
